Move cartesian impedance controller into the base robot
