+++ arm_robots/src/arm_robots/med.py
@@ -69,6 +69,7 @@ class BaseMed(BaseRobot):
         self.get_control_mode_srv = rospy.ServiceProxy(self.ns("get_control_mode_service"), GetControlMode)
         self.arm_status_listener = Listener(self.ns("motion_status"), MotionStatus)
         self.waypoint_state_pub = rospy.Publisher(self.ns("waypoint_robot_state"), DisplayRobotState, queue_size=10)
+        self.create_cartesian_impedance_controller([self.arm_status_listener], [self.arm_command_pub], ARM_JOINT_NAMES)
 
     def send_joint_command(self, joint_names: Sequence[str], trajectory_point: JointTrajectoryPoint) -> Tuple[
         bool, str]:
